Se agrego un controlador y el Service de Respuesta, falta añadir clases Ticket, Asignacion y Cliente

+++ backend-sqrc/src/main/java/com/sqrc/module/backendsqrc/plantillaRespuesta/model/RespuestaCliente.java
@@ -18,7 +18,7 @@ public class RespuestaCliente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_respuesta")
-    private Integer idRespuesta;
+    private Long idRespuesta;
 
 /*
     @ManyToOne(fetch = FetchType.LAZY)
